backend/python_solver/requirements.txt

fastapi>=0.100.0
uvicorn>=0.23.0
ortools>=9.7.0
tensorflow-cpu>=2.13.0
firebase-admin>=6.2.0
psutil>=5.9.0
python-multipart
requests>=2.31.0
google-cloud-storage
google-cloud-datastore
fpdf2
pydantic>=2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
google-genai>=0.3.0
google-cloud-tasks>=2.13.0
cachetools
